Switch name and add refactored fit function
